switch to a working branch before add, commit, push package updates

+++ pkg/update/update.go
@@ -7,6 +7,8 @@ import (
 	"path/filepath"
 	"time"
 
+	"github.com/go-git/go-git/v5/plumbing"
+
 	"github.com/hashicorp/go-version"
 
 	"github.com/go-git/go-git/v5"
@@ -33,7 +35,7 @@ func New() (Context, error) {
 			client: http.DefaultClient,
 
 			// 1 request every (n) second(s) to avoid DOS'ing server
-			Ratelimiter: rate.NewLimiter(rate.Every(1*time.Second), 5),
+			Ratelimiter: rate.NewLimiter(rate.Every(2*time.Second), 5),
 		},
 		Logger: log.New(log.Writer(), "wupdater: ", log.LstdFlags|log.Lmsgprefix),
 	}
@@ -50,10 +52,14 @@ func (c Context) Update() error {
 	if err != nil {
 		return errors.Wrapf(err, "failed to create temporary folder to clone package configs into")
 	}
-	c.Logger.Printf("using temp dir %s", tempDir)
-	defer os.Remove(tempDir)
 
-	_, err = git.PlainClone(tempDir, false, &git.CloneOptions{
+	if c.DryRun {
+		c.Logger.Printf("using temp dir %s", tempDir)
+	} else {
+		defer os.Remove(tempDir)
+	}
+
+	repo, err := git.PlainClone(tempDir, false, &git.CloneOptions{
 		URL:      c.RepoURI,
 		Progress: os.Stdout,
 	})
@@ -74,6 +80,7 @@ func (c Context) Update() error {
 		return errors.Wrapf(err, "failed getting release monitor service mapping data")
 	}
 
+	packagesToUpdate := make(map[string]string)
 	for _, config := range c.Packages {
 		item := mapperData[config.Package.Name]
 		if item.Identifier == "" {
@@ -94,16 +101,85 @@ func (c Context) Update() error {
 			c.Logger.Printf("failed to create a version from package %s: %s.  Error: %s", config.Package.Name, latestVersion, err)
 			continue
 		}
-		//c.Logger.Printf("package %s, latest available version %s, current version %s", config.Package.Name, config.Package.Version, latestVersion)
 
 		if currentVersionSemver.LessThan(latestVersionSemver) {
 			c.Logger.Printf("there is a new stable version available %s %s, current wolfi version %s", config.Package.Name, latestVersion, config.Package.Version)
+			packagesToUpdate[config.Package.Name] = latestVersion
 		}
 
-		// if new versions are available, create a pull request
+	}
+
+	err = c.updatePackagesGitRepository(repo, packagesToUpdate, tempDir)
+	if err != nil {
+		return errors.Wrapf(err, "failed to make updates on %s", c.RepoURI)
+	}
+
+	// diff any changes
+	worktree, err := repo.Worktree()
+	if err != nil {
+		return errors.Wrapf(err, "failed to get git worktree")
+	}
+
+	_, err = worktree.Commit("Wolfi update packages", &git.CommitOptions{})
+	if err != nil {
+		return errors.Wrapf(err, "failed to git commit")
+	}
+
+	if !c.DryRun {
+		err = repo.Push(&git.PushOptions{RemoteName: "origin"})
+		if err != nil {
+			return errors.Wrapf(err, "failed to git push")
+		}
 	}
 
 	return nil
+}
+
+func (c Context) updatePackagesGitRepository(repo *git.Repository, packagesToUpdate map[string]string, tempDir string) error {
+	// let's work on a branch when updating package versions, so we can create a PR from that branch later
+	err := c.switchBranch(repo)
+	if err != nil {
+		return errors.Wrapf(err, "failed to switch to working git branch")
+	}
+
+	// bump packages that need updating
+	for packageName, latestVersion := range packagesToUpdate {
+
+		// if new versions are available lets bump the packages in the target melange git repo
+		//if c.Batch {
+		configFile := filepath.Join(tempDir, packageName+".yaml")
+
+		err = c.bump(configFile, latestVersion)
+		if err != nil {
+			c.Logger.Printf("failed to bump config file %s to version %s: %s", configFile, latestVersion, err.Error())
+		}
+		worktree, err := repo.Worktree()
+		if err != nil {
+			return errors.Wrapf(err, "failed to get git worktree")
+		}
+		_, err = worktree.Add(packageName + ".yaml")
+		if err != nil {
+			return errors.Wrapf(err, "failed to get add %s", configFile)
+		}
+	}
+
+	return nil
+}
+
+// create a unique branch
+func (c Context) switchBranch(repo *git.Repository) error {
+	name := time.Now().Format("2006102150405")
+
+	worktree, err := repo.Worktree()
+	if err != nil {
+		return errors.Wrapf(err, "failed to get git worktree")
+	}
+	ref := plumbing.ReferenceName("refs/heads/wupdater-" + name)
+	err = worktree.Checkout(&git.CheckoutOptions{
+		Create: true,
+		Branch: ref,
+	})
+	return err
 }
 
 func (c Context) getPackageConfigs(tempDir string) error {
